Give authenticated pages their own layout hierarchy This means we don't need as many if user checks, which are tricky, because the user isn't always in sync with the page route (e.g. the user gets set before leaving the login page, but the login page layout should assume there's no user).
