+++ frontend/src/lib/layout/Content.svelte
@@ -1,0 +1,3 @@
+<main class="max-w-none px-2 md:px-6 pt-8">
+  <slot />
+</main>
