+++ frontend/src/lib/layout/index.ts
@@ -2,11 +2,13 @@ import AppBar from './AppBar.svelte'
 import AppMenu from './AppMenu.svelte'
 import Breadcrumbs from './Breadcrumbs.svelte'
 import Page from './Page.svelte'
+import Content from './Content.svelte'
 
 export {
   AppBar,
   AppMenu,
   Page,
   Breadcrumbs,
+  Content,
 }
