+++ frontend/src/routes/(authenticated)/(dashboards)/UserDashboard.svelte
