+++ frontend/src/routes/(authenticated)/(dashboards)/admin/+page.svelte
