+++ frontend/src/routes/(authenticated)/(dashboards)/admin/+page.ts
