+++ frontend/src/routes/(authenticated)/+page.svelte
