+++ frontend/src/routes/(authenticated)/+page.ts
