+++ frontend/src/routes/(authenticated)/logout/+server.ts
