+++ frontend/src/routes/(authenticated)/project/[project_code]/+page.svelte
