+++ frontend/src/routes/(authenticated)/project/create/+page.svelte
