+++ frontend/src/routes/(authenticated)/resetPassword/+page.svelte
