+++ frontend/src/routes/(unauthenticated)/+layout.svelte
@@ -1,0 +1,9 @@
+<script>
+  import { AppBar, Content } from '$lib/layout';
+</script>
+
+<AppBar />
+
+<Content>
+  <slot />
+</Content>
