+++ frontend/src/routes/(unauthenticated)/forgotPassword/+page.svelte
