+++ frontend/src/routes/(unauthenticated)/forgotPassword/emailSent/+page.svelte
