+++ frontend/src/routes/(unauthenticated)/register/+page.svelte
