+++ frontend/src/routes/+layout.svelte
@@ -3,24 +3,8 @@
   import '$lib/app.postcss';
   import { error } from '$lib/error';
   import UnexpectedErrorAlert from '$lib/error/UnexpectedErrorAlert.svelte';
-  import { AppBar, AppMenu, Breadcrumbs } from '$lib/layout';
-  import { user } from '$lib/user';
   import { onDestroy } from 'svelte';
   import type { LayoutData } from './$types';
-
-  let menuToggle = false;
-
-  function open(): void {
-    menuToggle = true;
-  }
-
-  function close(): void {
-    menuToggle = false;
-  }
-
-  function closeOnEscape(event: KeyboardEvent): void {
-    event.key === 'Escape' && close();
-  }
 
   onDestroy(
     page.subscribe((p) => {
@@ -33,8 +17,6 @@
   export let data: LayoutData;
 </script>
 
-<svelte:window on:keydown={closeOnEscape} />
-
 <svelte:head>
   {#if data.traceParent}
     <meta name="traceparent" content={data.traceParent} />
@@ -42,26 +24,8 @@
 </svelte:head>
 
 <!-- https://daisyui.com/components/drawer -->
-<div class="drawer drawer-end">
-  <input type="checkbox" checked={menuToggle} class="drawer-toggle" />
-
-  <div class="drawer-content">
-    <AppBar on:menuopen={open} />
-    {#if $user}
-      <div class="bg-secondary p-2 pl-6">
-        <Breadcrumbs />
-      </div>
-    {/if}
-
-    <!-- https://tailwindcss.com/docs/typography-plugin -->
-    <main class="max-w-none px-2 md:px-6 pt-8">
-      <slot />
-    </main>
-  </div>
-
-  {#if $user}
-    <AppMenu on:click={close} on:keydown={close} />
-  {/if}
+<div>
+  <slot />
 </div>
 
 <!-- We don't want the alert for "hook" errors, because they land on +error.svelte -->
